# GH-3523 Cache mechanism for MemStatementIterator(s)