… core/sail/memory/src/main/java/org/eclipse/rdf4j/sail/memory/model/MemStatementIteratorCache.java