@hmottestad 2021-12-17
### After
```
Benchmark                            Mode  Cnt   Score   Error  Units
QueryBenchmark.complexQuery          avgt    5   6.617 ± 0.020  ms/op
QueryBenchmark.groupByQuery          avgt    5   2.026 ± 0.016  ms/op
QueryBenchmark.pathExpressionQuery1  avgt    5  25.105 ± 0.418  ms/op
QueryBenchmark.pathExpressionQuery2  avgt    5   1.443 ± 0.060  ms/op
```