@hmottestad 2021-12-17
Also need to benchmark what cost this has for write operations.