@JervenBolleman 2021-12-17
What happens if to queries run in parallel and access the same iterator?