@hmottestad 2021-12-17
It's not the actual iterator that is cached, but an ArrayList with the results from the iterator. So if two threads both hit the cache then they will each get a new iterator backed by the list.