@hmottestad 2021-12-18
### Before
```
Benchmark                                                     Mode  Cnt      Score     Error  Units
QueryBenchmark.complexQuery                                   avgt    5     47.596 ±   0.488  ms/op
QueryBenchmark.different_datasets_with_similar_distributions  avgt    5      3.232 ±   0.090  ms/op
QueryBenchmark.groupByQuery                                   avgt    5      1.983 ±   0.072  ms/op
QueryBenchmark.long_chain                                     avgt    5   1055.210 ±  32.642  ms/op
QueryBenchmark.lots_of_optional                               avgt    5    271.959 ±   8.758  ms/op
QueryBenchmark.minus                                          avgt    5  12103.994 ± 100.582  ms/op
QueryBenchmark.nested_optionals                               avgt    5    262.800 ±   5.527  ms/op
QueryBenchmark.pathExpressionQuery1                           avgt    5     23.529 ±   0.894  ms/op
QueryBenchmark.pathExpressionQuery2                           avgt    5      1.596 ±   0.102  ms/op
QueryBenchmark.query_distinct_predicates                      avgt    5    194.093 ±   9.729  ms/op
QueryBenchmark.simple_filter_not                              avgt    5     10.962 ±   0.539  ms/op
```

### After
```
Benchmark                                                     Mode  Cnt      Score     Error  Units
QueryBenchmark.complexQuery                                   avgt    5      6.758 ±   0.084  ms/op
QueryBenchmark.different_datasets_with_similar_distributions  avgt    5      3.324 ±   0.128  ms/op
QueryBenchmark.groupByQuery                                   avgt    5      2.056 ±   0.077  ms/op
QueryBenchmark.long_chain                                     avgt    5   1044.142 ±  31.316  ms/op
QueryBenchmark.lots_of_optional                               avgt    5    282.401 ±   7.528  ms/op
QueryBenchmark.minus                                          avgt    5  12203.275 ± 410.164  ms/op
QueryBenchmark.nested_optionals                               avgt    5    268.526 ±   6.198  ms/op
QueryBenchmark.pathExpressionQuery1                           avgt    5     23.477 ±   0.752  ms/op
QueryBenchmark.pathExpressionQuery2                           avgt    5      1.553 ±   0.123  ms/op
QueryBenchmark.query_distinct_predicates                      avgt    5    169.904 ±   1.297  ms/op
QueryBenchmark.simple_filter_not                              avgt    5     10.334 ±   0.221  ms/op
```