@hmottestad 2021-12-18
So this doesn't really benefit the other queries much at the moment.