@hmottestad 2021-12-19
I'll go over the naming a bit. Maybe something like CacheAwareIteration would be more informative. Also need to check how much this impacts data insertion.