@hmottestad 2022-01-05
@jeenbroekstra Could you take another look? I've moved around some of the code and renamed things to make it all a bit more obvious what the code is supposed to do.